add classes csvtofirestore and olshoptransaction

+++ app/src/main/java/com/mrap/olshopman/Item.java
@@ -7,5 +7,5 @@ import java.util.ArrayDeque;
 public class Item {
     public String id = null;
     public String name;
-    public ArrayDeque<Pair<Item, Float>> materials = new ArrayDeque<>();
+    public ArrayDeque<Pair<Item, Double>> materials = new ArrayDeque<>();
 }
